+++ src/med_bench/mediation.py
@@ -9,7 +9,6 @@ causal inference, simulate data, and evaluate and compare estimators
 import numpy as np
 import pandas as pd
 from sklearn.base import clone
-from sklearn.linear_model import RidgeCV
 
 
 from .utils.nuisances import (
