+++ src/med_bench/utils/utils.py
@@ -7,6 +7,8 @@ from sklearn.model_selection import KFold
 
 import subprocess
 
+from med_bench.utils.constants import ALPHAS, TINY
+
 
 def check_r_dependencies():
     try:
@@ -35,6 +37,58 @@ def check_r_dependencies():
     except:
         # Handle the case where R is not found or rpy2 is not installed
         return False
+
+
+def _get_interactions(interaction, *args):
+    """
+    this function provides interaction terms between different groups of
+    variables (confounders, treatment, mediators)
+
+    Parameters
+    ----------
+    interaction : boolean
+                    whether to compute interaction terms
+
+    *args : flexible, one or several arrays
+                    blocks of variables between which interactions should be
+                    computed
+
+
+    Returns
+    --------
+    array_like
+        interaction terms
+
+    Examples
+    --------
+    >>> x = np.arange(6).reshape(3, 2)
+    >>> t = np.ones((3, 1))
+    >>> m = 2 * np.ones((3, 1))
+    >>> get_interactions(False, x, t, m)
+    array([[0., 1., 1., 2.],
+           [2., 3., 1., 2.],
+           [4., 5., 1., 2.]])
+    >>> get_interactions(True, x, t, m)
+    array([[ 0.,  1.,  1.,  2.,  0.,  1.,  0.,  2.,  2.],
+           [ 2.,  3.,  1.,  2.,  2.,  3.,  4.,  6.,  2.],
+           [ 4.,  5.,  1.,  2.,  4.,  5.,  8., 10.,  2.]])
+    """
+    variables = list(args)
+    for index, var in enumerate(variables):
+        if len(var.shape) == 1:
+            variables[index] = var.reshape(-1, 1)
+    pre_inter_variables = np.hstack(variables)
+    if not interaction:
+        return pre_inter_variables
+    new_cols = list()
+    for i, var in enumerate(variables[:]):
+        for j, var2 in enumerate(variables[i + 1:]):
+            for ii in range(var.shape[1]):
+                for jj in range(var2.shape[1]):
+                    new_cols.append((var[:, ii] * var2[:, jj]).reshape(-1, 1))
+    new_vars = np.hstack(new_cols)
+    result = np.hstack((pre_inter_variables, new_vars))
+    return result
 
 
 def is_r_installed():
@@ -104,58 +158,6 @@ def r_dependency_required(required_packages):
 
 if is_r_installed():
     import rpy2.robjects as robjects
-
-
-def _get_interactions(interaction, *args):
-    """
-    this function provides interaction terms between different groups of
-    variables (confounders, treatment, mediators)
-
-    Parameters
-    ----------
-    interaction : boolean
-                    whether to compute interaction terms
-
-    *args : flexible, one or several arrays
-                    blocks of variables between which interactions should be
-                    computed
-
-
-    Returns
-    --------
-    array_like
-        interaction terms
-
-    Examples
-    --------
-    >>> x = np.arange(6).reshape(3, 2)
-    >>> t = np.ones((3, 1))
-    >>> m = 2 * np.ones((3, 1))
-    >>> get_interactions(False, x, t, m)
-    array([[0., 1., 1., 2.],
-           [2., 3., 1., 2.],
-           [4., 5., 1., 2.]])
-    >>> get_interactions(True, x, t, m)
-    array([[ 0.,  1.,  1.,  2.,  0.,  1.,  0.,  2.,  2.],
-           [ 2.,  3.,  1.,  2.,  2.,  3.,  4.,  6.,  2.],
-           [ 4.,  5.,  1.,  2.,  4.,  5.,  8., 10.,  2.]])
-    """
-    variables = list(args)
-    for index, var in enumerate(variables):
-        if len(var.shape) == 1:
-            variables[index] = var.reshape(-1, 1)
-    pre_inter_variables = np.hstack(variables)
-    if not interaction:
-        return pre_inter_variables
-    new_cols = list()
-    for i, var in enumerate(variables[:]):
-        for j, var2 in enumerate(variables[i + 1 :]):
-            for ii in range(var.shape[1]):
-                for jj in range(var2.shape[1]):
-                    new_cols.append((var[:, ii] * var2[:, jj]).reshape(-1, 1))
-    new_vars = np.hstack(new_cols)
-    result = np.hstack((pre_inter_variables, new_vars))
-    return result
 
 
 def _convert_array_to_R(x):
@@ -244,7 +246,8 @@ def _check_input(y, t, m, x, setting):
         raise ValueError("Multidimensional m (mediator) is not supported")
 
     if (setting == "binary") and (len(np.unique(m)) != 2):
-        raise ValueError("Only a binary one-dimensional m (mediator) is supported")
+        raise ValueError(
+            "Only a binary one-dimensional m (mediator) is supported")
 
     return y_converted, t_converted, m_converted, x_converted
 
@@ -255,48 +258,23 @@ def is_array_integer(array):
     return all(list((array == array.astype(int)).squeeze()))
 
 
-def str_to_bool(string):
-    if bool(string) == string:
-        return string
-    elif string == "True":
-        return True
-    elif string == "False":
-        return False
-    else:
-        raise ValueError  # evil ValueError that doesn't tell you what the wrong value was
-
-
-def bucketize_mediators(m, n_buckets=10, random_state=42):
-    kmeans = KMeans(n_clusters=n_buckets, random_state=random_state, n_init="auto").fit(
-        m
-    )
-    return kmeans.predict(m)
-
-
-def train_test_split_data(causal_data, test_size=0.33, random_state=42):
-    x, t, m, y = causal_data
-    x_train, x_test, t_train, t_test, m_train, m_test, y_train, y_test = (
-        train_test_split(x, t, m, y, test_size=test_size, random_state=random_state)
-    )
-    causal_data_train = x_train, t_train, m_train, y_train
-    causal_data_test = x_test, t_test, m_test, y_test
-    return causal_data_train, causal_data_test
-
-
-def _get_train_test_lists(crossfit, n, x):
+def _get_regularization_parameters(regularization):
     """
-    Obtain train and test folds
+    Obtain regularization parameters
 
     Returns
     -------
-    train_test_list : list
-        indexes with train and test indexes
+    cs : list
+        each of the values in Cs describes the inverse of regularization
+        strength for predictors
+    alphas : list
+        alpha values to try in ridge models
     """
-    if crossfit < 2:
-        train_test_list = [[np.arange(n), np.arange(n)]]
+    if regularization:
+        alphas = ALPHAS
+        cs = ALPHAS
     else:
-        kf = KFold(n_splits=crossfit)
-        train_test_list = list()
-        for train_index, test_index in kf.split(x):
-            train_test_list.append([train_index, test_index])
-    return train_test_list
+        alphas = [TINY]
+        cs = [np.inf]
+
+    return cs, alphas
